feat: deterministic md recomposition order
* feat: deterministic md recomposition order

* chore: sort order for recompose

* test: deep equal in any order

* test: order of __c vs letters

* test: underscores first

* test: more underscore first

* chore: docs and export cleanup

* refactor: avoid unnecessary compSet

* chore: console.log cleanup
